# swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Schools",
    "version": "1.0.0",
    "title": "Schools API"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/schools": {
      "get": {
        "summary": "Get all schools",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/School"
            }
          }
        }
      },
      "post": {
        "summary": "Add a new school",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "School object that needs to be added to the store",
            "required": true,
            "schema": {
              "$ref": "#/definitions/School"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/schools/{schoolId}": {
      "get": {
        "summary": "Find school by ID",
        "description": "Returns a single school",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "schoolId",
            "in": "path",
            "description": "ID of school to return",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/School"
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "School not found"
          }
        }
      }
    }
  },
  "definitions": {
    "School": {
      "type": "object",
      "required": [
        "name"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string",
          "example": "School-123"
        }
      }
    }
  }
}